chore(wmg): move Heatmap assets
