+++ frontend/src/views/WheresMyGeneV2/components/HeatMap/hooks/useSortedCellTypesByTissueName.ts
