+++ frontend/src/views/WheresMyGeneV2/components/HeatMap/index.tsx
@@ -40,8 +40,8 @@ import YAxisChart from "./components/YAxisChart";
 import {
   useSortedGeneNames,
   useTissueNameToCellTypeIdToGeneNameToCellTypeGeneExpressionSummaryDataMap,
-} from "src/views/WheresMyGene/components/HeatMap/hooks/useSortedGeneNames";
-import { useSortedCellTypesByTissueName } from "src/views/WheresMyGene/components/HeatMap/hooks/useSortedCellTypesByTissueName";
+} from "src/views/WheresMyGeneV2/components/HeatMap/hooks/useSortedGeneNames";
+import { useSortedCellTypesByTissueName } from "src/views/WheresMyGeneV2/components/HeatMap/hooks/useSortedCellTypesByTissueName";
 import {
   CellTypeTagContainer,
   ChartWrapper,
@@ -50,7 +50,7 @@ import {
   StyledTag,
   XAxisMask,
   YAxisWrapper,
-} from "src/views/WheresMyGene/components/HeatMap/style";
+} from "src/views/WheresMyGeneV2/components/HeatMap/style_old";
 import { CellCountLabel } from "src/views/WheresMyGeneV2/components/HeatMap/components/XAxisChart/style";
 import {
   HEATMAP_CONTAINER_ID,
@@ -59,7 +59,7 @@ import {
 import Loader from "src/views/WheresMyGene/components/Loader";
 import XAxisChart from "src/views/WheresMyGeneV2/components/HeatMap/components/XAxisChart";
 import Chart from "src/views/WheresMyGeneV2/components/HeatMap/components/Chart";
-import { hyphenize } from "src/views/WheresMyGene/components/HeatMap/utils";
+import { hyphenize } from "src/views/WheresMyGeneV2/components/HeatMap/utils";
 import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "../GeneSearchBar/components/SaveExport";
 import { Autocomplete, DefaultAutocompleteOption } from "@czi-sds/components";
 import {
