+++ frontend/src/views/WheresMyGeneV2/components/HeatMap/style_old.ts
@@ -1,14 +1,14 @@
 import styled from "@emotion/styled";
 import { Y_AXIS_CHART_WIDTH_PX } from "./utils";
 import { LIGHT_GRAY } from "src/components/common/theme";
-import { LEGEND_HEIGHT_PX } from "../InfoPanel/components/Legend/style";
+import { LEGEND_HEIGHT_PX } from "../../../WheresMyGene/components/InfoPanel/components/Legend/style";
 import { HEADER_HEIGHT_PX } from "src/components/Header/style";
 import {
   CONTENT_WRAPPER_LEFT_RIGHT_PADDING_PX,
   CONTENT_WRAPPER_TOP_BOTTOM_PADDING_PX,
 } from "src/components/Layout/style";
-import { LEGEND_MARGIN_BOTTOM_PX } from "../../style";
-import { X_AXIS_CHART_HEIGHT_PX } from "../../../WheresMyGeneV2/common/constants";
+import { LEGEND_MARGIN_BOTTOM_PX } from "../../../WheresMyGene/style";
+import { X_AXIS_CHART_HEIGHT_PX } from "../../common/constants";
 import { Autocomplete } from "@mui/material";
 import { TagFilter } from "@czi-sds/components";
 import { spacesS } from "src/common/theme";
